pyproject.toml
[project]
name = "topicops"
version = "0.1.0"
description = "Open-source control plane for collection intent across scrapers, OSINT workflows, feeds, and agent-accessible intelligence systems."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "TopicOps contributors" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.0",
  "rich>=13.7.0",
  "httpx>=0.27.0",
  "feedparser>=6.0.11",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.0",
  "jsonschema>=4.22.0",
  "rapidfuzz>=3.9.0",
  "networkx>=3.3.0",
  "semver>=3.0.2",
  "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12",
]

[project.scripts]
topicops = "topicops.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
